Created a "See all" button as the last tile of the list and added a trailing arrow icon.
Notes :
According to the current design, the screen is not scrollable, so if See all is clicked, the user should be navigated to a new screen with the full list. If the whole list should be displayed in this screen itself, then the container right after the AppBar should be replaced with something scrollable.